them modal thong bao khi xoa

+++ src/components/templates/DatPhong/DatPhongList.tsx
@@ -12,6 +12,7 @@ import {
 import { EditDatPhong } from "./EditDatPhong";
 import { toast } from "react-toastify";
 import { handleError } from "utils";
+import Swal from "sweetalert2";
 export const DatPhongList = () => {
   const dispatch = useAppDispatch();
 
@@ -76,15 +77,26 @@ export const DatPhongList = () => {
           <Button
             className="mr-[15px] !bg-red-600 !text-white !font-500 "
             onClick={() => {
-              dispatch(DeleteDatPhongThunk(record.key))
-                .unwrap()
-                .then(() => {
-                  try {
-                    toast.success("Xóa thành công");
-                  } catch (error) {
-                    handleError(error);
-                  }
-                });
+              Swal.fire({
+                title: "Bạn có muốn xóa không",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes",
+              }).then((result) => {
+                if (result.isConfirmed) {
+                  dispatch(DeleteDatPhongThunk(record.key))
+                    .unwrap()
+                    .then(() => {
+                      try {
+                        toast.success("Xóa thành công");
+                      } catch (error) {
+                        handleError(error);
+                      }
+                    });
+                }
+              });
             }}
           >
             <i className="fa-solid fa-trash"></i>
