+++ src/components/templates/Room/RoomDS.tsx
@@ -10,6 +10,8 @@ import { RoomListService } from "services";
 import { UploadOutlined } from "@ant-design/icons";
 import { handleError } from "utils";
 import { toast } from "react-toastify";
+import Swal from "sweetalert2";
+
 export const RoomDS = () => {
   const { roomList } = useSelector((state: RootState) => state.RoomReducer);
   const dispatch = useAppDispatch();
@@ -91,14 +93,26 @@ export const RoomDS = () => {
           <Button
             className="mr-[15px] !bg-red-600 !text-white !font-500 "
             onClick={() => {
-              dispatch(DeleteRoomThunk(record.key))
-                .unwrap()
-                .then(() => {
-                  toast.success("Xóa thành công");
-                })
-                .catch((error) => {
-                  handleError(error);
-                });
+              dispatch(DeleteRoomThunk(record.key));
+              Swal.fire({
+                title: "Bạn có muốn xóa không",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes",
+              }).then((result) => {
+                if (result.isConfirmed) {
+                  dispatch(DeleteRoomThunk(record.key))
+                    .unwrap()
+                    .then(() => {
+                      toast.success("Xóa thành công");
+                    })
+                    .catch((error) => {
+                      handleError(error);
+                    });
+                }
+              });
             }}
           >
             <i className="fa-solid fa-trash"></i>
